// packages/core/src/types/bot.d.ts
declare namespace BotModel {
// ...
  interface Token {
    id: number;
    code: string;
    symbol: string;
    name: string;
    icon: string;
    decimals: number;
    oracle_id: string;
    volume: string;
    amount?: string;
  }
// ...
}